+++ tests/unit/core/test_etl_dataframe.py
@@ -322,6 +322,7 @@ def test_pool_with_func_returning_series_with_different_shape(dataframe1):
     )
 
 
+@pytest.mark.parametrize("as_type", ["dict", "list", "kwargs"])
 @pytest.mark.parametrize(
     "query, expected",
     [
@@ -375,18 +376,38 @@ def test_pool_with_func_returning_series_with_different_shape(dataframe1):
         ),
     ],
 )
-def test_query(dataframe1, query, expected):
+def test_query(dataframe1, query, expected, as_type):
     obj = dataframe1
+    assert isinstance(query, dict)
 
-    # test query with dict
+    if as_type == "dict":
+        result = obj.etl.q(query)
+    elif as_type == "list":
+        result = obj.etl.q([query])
+    elif as_type == "kwargs":
+        result = obj.etl.q(**query)
+    else:
+        raise ValueError(f"Invalid as_type: {as_type}")
+
+    # check_index_type=False needed to avoid Attribute "inferred_type" are different
+    assert_frame_equal(result, expected, check_index_type=False)
+
+
+def test_query_with_multiple_dicts(dataframe1):
+    obj = dataframe1
+    query = [
+        {"v0": 0},
+        {"v0": 1, "v1": 5},
+    ]
+    expected = pd.DataFrame(
+        [[0, 4], [1, 5]],
+        columns=["v0", "v1"],
+        index=pd.MultiIndex.from_tuples([("a", "c"), ("a", "d")], names=["i0", "i1"]),
+    )
+
     result = obj.etl.q(query)
-    # check_index_type=False needed to avoid Attribute "inferred_type" are different
-    assert_frame_equal(result, expected, check_index_type=False)
 
-    # test query with params
-    result = obj.etl.q(**query)
-    # check_index_type=False needed to avoid Attribute "inferred_type" are different
-    assert_frame_equal(result, expected, check_index_type=False)
+    assert_frame_equal(result, expected)
 
 
 def test_query_with_invalid_params(dataframe1):